+++ unravel/cfg.py
@@ -33,7 +33,8 @@ CUR_STAGE = 0
 USE_BOX_CHECK_V1 = False
 
 
-@jit(nopython=True)
+
+@jit(nopython=True, cache=True)
 def log(*args) -> None:
     if SHOW_PROGRESS:
         print(*args)